+++ MicroServices/BonAppetit.RestaurantServices/Models/ScheduleModels/ScheduleBase.cs
@@ -8,26 +8,26 @@ public class ScheduleBase
 
     public string ScheduleId { get; set; } = Guid.NewGuid().ToString();
     public bool Sunday { get; set; }
-    public int SundayOpenTime { get; set; }
-    public int SundayCloseTime { get; set; }
+    public double SundayOpenTime { get; set; }
+    public double SundayCloseTime { get; set; }
     public bool Monday { get; set; }
-    public int MondayOpenTime { get; set; }
-    public int MondayCloseTime { get; set; }
+    public double MondayOpenTime { get; set; }
+    public double MondayCloseTime { get; set; }
     public bool Tuesday { get; set; }
-    public int TuesdayOpenTime { get; set; }
-    public int TuesdayCloseTime { get; set; }
+    public double TuesdayOpenTime { get; set; }
+    public double TuesdayCloseTime { get; set; }
     public bool Wednesday { get; set; }
-    public int WednesdayOpenTime { get; set; }
-    public int WednesdayCloseTime { get; set; }
+    public double WednesdayOpenTime { get; set; }
+    public double WednesdayCloseTime { get; set; }
     public bool Thursday { get; set; }
-    public int ThursdayOpenTime { get; set; }
-    public int ThursdayCloseTime { get; set; }
+    public double ThursdayOpenTime { get; set; }
+    public double ThursdayCloseTime { get; set; }
     public bool Friday { get; set; }
-    public int FridayOpenTime { get; set; }
-    public int FridayCloseTime { get; set; }
+    public double FridayOpenTime { get; set; }
+    public double FridayCloseTime { get; set; }
     public bool Saturday { get; set; }
-    public int SaturdayOpenTime { get; set; }
-    public int SaturdayCloseTime { get; set; }
+    public double SaturdayOpenTime { get; set; }
+    public double SaturdayCloseTime { get; set; }
 
     public RestaurantBase Restaurant { get; set; }
     public string RestaurantId { get; set; }
