+++ MicroServices/BonAppetit.RestaurantServices/Models/ScheduleModels/ScheduleDto.cs
@@ -17,64 +17,64 @@ public class ScheduleDto
     public bool Sunday { get; set; }
 
     [Required(AllowEmptyStrings = false)]
-    public int SundayOpenTime { get; set; }
+    public double SundayOpenTime { get; set; }
 
     [Required(AllowEmptyStrings = false)]
-    public int SundayCloseTime { get; set; }
+    public double SundayCloseTime { get; set; }
 
     [Required(AllowEmptyStrings = false)]
     public bool Monday { get; set; }
 
     [Required(AllowEmptyStrings = false)]
-    public int MondayOpenTime { get; set; }
+    public double MondayOpenTime { get; set; }
 
     [Required(AllowEmptyStrings = false)]
-    public int MondayCloseTime { get; set; }
+    public double MondayCloseTime { get; set; }
 
     [Required(AllowEmptyStrings = false)]
     public bool Tuesday { get; set; }
 
     [Required(AllowEmptyStrings = false)]
-    public int TuesdayOpenTime { get; set; }
+    public double TuesdayOpenTime { get; set; }
 
     [Required(AllowEmptyStrings = false)]
-    public int TuesdayCloseTime { get; set; }
+    public double TuesdayCloseTime { get; set; }
 
     [Required(AllowEmptyStrings = false)]
     public bool Wednesday { get; set; }
 
     [Required(AllowEmptyStrings = false)]
-    public int WednesdayOpenTime { get; set; }
+    public double WednesdayOpenTime { get; set; }
 
     [Required(AllowEmptyStrings = false)]
-    public int WednesdayCloseTime { get; set; }
+    public double WednesdayCloseTime { get; set; }
 
     [Required(AllowEmptyStrings = false)]
     public bool Thursday { get; set; }
 
     [Required(AllowEmptyStrings = false)]
-    public int ThursdayOpenTime { get; set; }
+    public double ThursdayOpenTime { get; set; }
 
     [Required(AllowEmptyStrings = false)]
-    public int ThursdayCloseTime { get; set; }
+    public double ThursdayCloseTime { get; set; }
 
     [Required(AllowEmptyStrings = false)]
     public bool Friday { get; set; }
 
     [Required(AllowEmptyStrings = false)]
-    public int FridayOpenTime { get; set; }
+    public double FridayOpenTime { get; set; }
 
     [Required(AllowEmptyStrings = false)]
-    public int FridayCloseTime { get; set; }
+    public double FridayCloseTime { get; set; }
 
     [Required(AllowEmptyStrings = false)]
     public bool Saturday { get; set; }
 
     [Required(AllowEmptyStrings = false)]
-    public int SaturdayOpenTime { get; set; }
+    public double SaturdayOpenTime { get; set; }
 
     [Required(AllowEmptyStrings = false)]
-    public int SaturdayCloseTime { get; set; }
+    public double SaturdayCloseTime { get; set; }
     public RestaurantDto Restaurant { get; set; }
     
     #endregion
