+++ MicroServices/BonAppetit.RestaurantServices/Models/TableModels/TableCreate.cs
@@ -9,7 +9,7 @@ public class TableCreate
     public string TableName { get; set; }
 
     [Required(AllowEmptyStrings = false)]
-    public int FrequencyOfReservation { get; set; }
+    public double FrequencyOfReservation { get; set; }
 
     [Required(AllowEmptyStrings = false)]
     public int AmountOfSeats { get; set; }
