+++ MicroServices/BonAppetit.RestaurantServices/Models/TableModels/TableDto.cs
@@ -13,7 +13,7 @@ public class TableDto
     public string TableName { get; set; }
 
     [Required(AllowEmptyStrings = false)]
-    public int FrequencyOfReservation { get; set; }
+    public double FrequencyOfReservation { get; set; }
 
     [Required(AllowEmptyStrings = false)]
     public int AmountOfSeats { get; set; }
